feat(a2ui): implement ElmA2uiSurfaceRenderer for rendering A2UI surfaces
